+++ NEWS.md
@@ -22,6 +22,7 @@ Active logins are now necessary to use webservices such as CSI:FingerID, see e.g
 - `delete()` function for MS peak lists
 - Fixed: `EICOnlyPresent` argument to `reportHTML()` is effective again
 - Fixed: blank filter didn't properly handle differing blank assignments per analysis
+- `generateTPsLibrary`: the `matchParentsBy` argument now also accepts `"formula"` and `"name"`.
 
 
 # patRoon 2.1